+++ cli/internal/metadataobject/graphql_schema_introspection/graphql_schema_introspection.go
@@ -48,7 +48,7 @@ type graphQLSchemaIntrospectionObject struct {
 }
 
 func (o *MetadataObject) Build() (map[string]interface{}, metadataobject.ErrParsingMetadataObject) {
-	data, err := ioutil.ReadFile(filepath.Join(o.MetadataDir, o.Filename()))
+	data, err := metadataobject.ReadMetadataFile(filepath.Join(o.MetadataDir, o.Filename()))
 	if err != nil {
 		return nil, o.error(err)
 	}